packagingv3: Move evaluate_simple_js_expression from configpanels to a separate utils file
