+++ src/utils/form.py
@@ -59,170 +59,15 @@ from pydantic.networks import EmailStr, HttpUrl
 from pydantic.types import constr
 
 from ..log import OperationLogger
-from ..utils.error import YunohostError, YunohostValidationError
-from ..utils.i18n import _value_for_locale
+from .error import YunohostError, YunohostValidationError
+from .i18n import _value_for_locale
 from .file_utils import read_yaml, write_to_file
+from .eval import evaluate_simple_js_expression
 
 if TYPE_CHECKING:
     from pydantic.fields import FieldInfo, ModelField
 
 logger = getLogger("yunohost.form")
-
-
-# ╭───────────────────────────────────────────────────────╮
-# │  ┌─╴╷ ╷╭─┐╷                                           │
-# │  ├─╴│╭╯├─┤│                                           │
-# │  ╰─╴╰╯ ╵ ╵╰─╴                                         │
-# ╰───────────────────────────────────────────────────────╯
-
-
-# Those js-like evaluate functions are used to eval safely visible attributes
-# The goal is to evaluate in the same way than js simple-evaluate
-# https://github.com/shepherdwind/simple-evaluate
-def evaluate_simple_ast(
-    node: ast.Expression | ast.expr, context: dict[str, Any] | None = None
-) -> Any:
-    if context is None:
-        context = {}
-
-    operators: dict[type, Callable[..., Any]] = {
-        ast.Not: op.not_,
-        ast.Mult: op.mul,
-        ast.Div: op.truediv,  # number
-        ast.Mod: op.mod,  # number
-        ast.Add: op.add,  # str
-        ast.Sub: op.sub,  # number
-        ast.USub: op.neg,  # Negative number
-        ast.Gt: op.gt,
-        ast.Lt: op.lt,
-        ast.GtE: op.ge,
-        ast.LtE: op.le,
-        ast.Eq: op.eq,
-        ast.NotEq: op.ne,
-    }
-    context["true"] = True
-    context["false"] = False
-    context["null"] = None
-
-    # Variable
-    if isinstance(node, ast.Name):  # Variable
-        return context[node.id]
-
-    # Python <=3.7 String
-    elif isinstance(node, ast.Str):
-        return node.s
-
-    # Python <=3.7 Number
-    elif isinstance(node, ast.Num):
-        return node.n
-
-    # Boolean, None and Python 3.8 for Number, Boolean, String and None
-    elif isinstance(node, (ast.Constant, ast.NameConstant)):
-        return node.value
-
-    # + - * / %
-    elif (
-        isinstance(node, ast.BinOp) and type(node.op) in operators
-    ):  # <left> <operator> <right>
-        left = evaluate_simple_ast(node.left, context)
-        right = evaluate_simple_ast(node.right, context)
-        if type(node.op) is ast.Add:
-            if isinstance(left, str) or isinstance(right, str):  # support 'I am ' + 42
-                left = str(left)
-                right = str(right)
-        elif type(left) is type(right):  # support "111" - "1" -> 110
-            left = float(left)
-            right = float(right)
-
-        return operators[type(node.op)](left, right)
-
-    # Comparison
-    # JS and Python don't give the same result for multi operators
-    # like True == 10 > 2.
-    elif (
-        isinstance(node, ast.Compare) and len(node.comparators) == 1
-    ):  # <left> <ops> <comparators>
-        left = evaluate_simple_ast(node.left, context)
-        right = evaluate_simple_ast(node.comparators[0], context)
-        operator = node.ops[0]
-        if isinstance(left, (int, float)) or isinstance(right, (int, float)):
-            try:
-                left = float(left)
-                right = float(right)
-            except ValueError:
-                return type(operator) is ast.NotEq
-        try:
-            return operators[type(operator)](left, right)
-        except TypeError:  # support "e" > 1 -> False like in JS
-            return False
-
-    # and / or
-    elif isinstance(node, ast.BoolOp):  # <op> <values>
-        for value in node.values:
-            value = evaluate_simple_ast(value, context)
-            if isinstance(node.op, ast.And) and not value:
-                return False
-            elif isinstance(node.op, ast.Or) and value:
-                return True
-        return isinstance(node.op, ast.And)
-
-    # not / USub (it's negation number -\d)
-    elif isinstance(node, ast.UnaryOp):  # <operator> <operand> e.g., -1
-        return operators[type(node.op)](evaluate_simple_ast(node.operand, context))
-
-    # match function call
-    elif isinstance(node, ast.Call) and node.func.__dict__.get("id") == "match":
-        return re.match(
-            evaluate_simple_ast(node.args[1], context),
-            context[node.args[0].id],  # type: ignore
-        )
-
-    # Unauthorized opcode
-    else:
-        opcode = str(type(node))
-        raise YunohostError(
-            f"Unauthorize opcode '{opcode}' in visible attribute", raw_msg=True
-        )
-
-
-def js_to_python(expr: str) -> str:
-    in_string = None
-    py_expr = ""
-    i = 0
-    escaped = False
-    for char in expr:
-        if char in r"\"'":
-            # Start a string
-            if not in_string:
-                in_string = char
-
-            # Finish a string
-            elif in_string == char and not escaped:
-                in_string = None
-
-        # If we are not in a string, replace operators
-        elif not in_string:
-            if char == "!" and expr[i + 1] != "=":
-                char = "not "
-            elif char in "|&" and py_expr[-1:] == char:
-                py_expr = py_expr[:-1]
-                char = " and " if char == "&" else " or "
-
-        # Determine if next loop will be in escaped mode
-        escaped = char == "\\" and not escaped
-        py_expr += char
-        i += 1
-    return py_expr
-
-
-def evaluate_simple_js_expression(
-    expr: str, context: Mapping[str, Any] = {}
-) -> int | float | str | bool | None | re.Match[str]:
-    if not expr.strip():
-        return False
-    node = ast.parse(js_to_python(expr), mode="eval").body
-    return evaluate_simple_ast(node, context)  # type: ignore
-
 
 # ╭───────────────────────────────────────────────────────╮
 # │  ╭─╮┌─╮╶┬╴╶┬╴╭─╮╭╮╷╭─╴                                │
@@ -866,7 +711,7 @@ class PasswordOption(BaseInputOption):
                 )
 
             # If it's an optional argument the value should be empty or strong enough
-            from ..utils.password import assert_password_is_strong_enough
+            from .password import assert_password_is_strong_enough
 
             assert_password_is_strong_enough("user", value)
